# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Klard – Booking platform for consultants
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  providerCount?: number;
  requiresDirectBilling?: boolean;
}

export type ProviderSummarySubscriptionTier =
  (typeof ProviderSummarySubscriptionTier)[keyof typeof ProviderSummarySubscriptionTier];

export const ProviderSummarySubscriptionTier = {
  basic: "basic",
  premium: "premium",
} as const;

export interface ProviderSummary {
  id: number;
  displayName: string;
  category: string;
  categorySlug?: string;
  city: string;
  zip: string;
  rating: number;
  reviewCount: number;
  minPrice: number;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  yearsExperience?: number | null;
  verified?: boolean;
  subscriptionTier?: ProviderSummarySubscriptionTier;
  requiresDirectBilling?: boolean;
}

export type ProviderSubscriptionTier =
  (typeof ProviderSubscriptionTier)[keyof typeof ProviderSubscriptionTier];

export const ProviderSubscriptionTier = {
  basic: "basic",
  premium: "premium",
} as const;

export interface Provider {
  id: number;
  clerkUserId: string;
  displayName: string;
  email?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bio?: string | null;
  category: string;
  categorySlug?: string;
  city: string;
  zip: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  yearsExperience?: number | null;
  rating: number;
  reviewCount: number;
  verified?: boolean;
  subscriptionTier?: ProviderSubscriptionTier;
  requiresDirectBilling?: boolean;
  /** @nullable */
  premiumSince?: string | null;
  /** @nullable */
  icalToken?: string | null;
  /** @nullable */
  calendarSyncUrl?: string | null;
  createdAt?: string;
}

export interface ProviderInput {
  displayName: string;
  bio?: string;
  category: string;
  city: string;
  zip: string;
  address?: string;
  phone?: string;
  website?: string;
  logoUrl?: string;
  yearsExperience?: number;
}

export interface ProviderUpdate {
  displayName?: string;
  bio?: string;
  category?: string;
  city?: string;
  zip?: string;
  address?: string;
  phone?: string;
  website?: string;
  avatarUrl?: string;
  logoUrl?: string;
  yearsExperience?: number;
}

export interface Service {
  id: number;
  providerId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  netPrice?: number | null;
  vatRate: number;
  durationMinutes: number;
  createdAt?: string;
}

export interface ServiceInput {
  name: string;
  description?: string;
  price?: number;
  netPrice?: number;
  vatRate?: number;
  durationMinutes: number;
}

export interface ServiceUpdate {
  name?: string;
  description?: string;
  price?: number;
  netPrice?: number;
  vatRate?: number;
  durationMinutes?: number;
}

export interface ServiceTemplate {
  id: number;
  categorySlug: string;
  name: string;
  /** @nullable */
  description?: string | null;
  defaultDurationMinutes: number;
  /** @nullable */
  defaultPrice?: number | null;
  sortOrder?: number;
}

export interface UploadUrlRequest {
  name: string;
  size?: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface TimeSlot {
  id: number;
  providerId: number;
  startTime: string;
  endTime: string;
  isAvailable: boolean;
}

export interface TimeSlotInput {
  providerId: number;
  startTime: string;
  endTime: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export type BookingPaymentStatus =
  (typeof BookingPaymentStatus)[keyof typeof BookingPaymentStatus];

export const BookingPaymentStatus = {
  not_required: "not_required",
  pending: "pending",
  paid: "paid",
  failed: "failed",
  refunded: "refunded",
} as const;

export interface Booking {
  id: number;
  customerId: string;
  /** @nullable */
  customerName?: string | null;
  providerId: number;
  providerName?: string;
  serviceId: number;
  serviceName?: string;
  slotId: number;
  status: BookingStatus;
  totalPrice: number;
  scheduledAt: string;
  /** @nullable */
  notes?: string | null;
  paymentRequired?: boolean;
  paymentStatus?: BookingPaymentStatus;
  /** @nullable */
  stripeCheckoutSessionId?: string | null;
  createdAt?: string;
}

export interface BookingInput {
  providerId: number;
  serviceId: number;
  slotId: number;
  notes?: string;
}

export type BookingStatusUpdateStatus =
  (typeof BookingStatusUpdateStatus)[keyof typeof BookingStatusUpdateStatus];

export const BookingStatusUpdateStatus = {
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
}

export interface Review {
  id: number;
  bookingId: number;
  customerId: string;
  /** @nullable */
  customerName?: string | null;
  providerId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  bookingId: number;
  rating: number;
  comment?: string;
}

export interface AiOfferRequest {
  providerId: number;
  inquiry: string;
}

export interface AiOfferResponse {
  offer: string;
  /** @nullable */
  estimatedPrice?: number | null;
  /** @nullable */
  estimatedDuration?: string | null;
}

export interface ProviderDashboard {
  totalBookings: number;
  pendingBookings: number;
  confirmedBookings: number;
  completedBookings: number;
  totalRevenue: number;
  averageRating: number;
  recentBookings: Booking[];
}

export interface CheckoutSession {
  url: string;
  sessionId?: string;
}

export type SubscriptionStatusTier =
  (typeof SubscriptionStatusTier)[keyof typeof SubscriptionStatusTier];

export const SubscriptionStatusTier = {
  basic: "basic",
  premium: "premium",
} as const;

export interface SubscriptionStatus {
  tier: SubscriptionStatusTier;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  currentPeriodEnd?: string | null;
  cancelAtPeriodEnd?: boolean;
  priceEur?: number;
}

export interface PlatformStats {
  totalProviders: number;
  totalBookings: number;
  totalCategories: number;
  totalReviews: number;
}

export type ListServiceTemplatesParams = {
  category: string;
};

export type ListProvidersParams = {
  zip?: string;
  city?: string;
  category?: string;
  q?: string;
  minPrice?: number;
  maxPrice?: number;
  limit?: number;
  offset?: number;
};

export type DeleteTimeSlotParams = {
  slotId: number;
};