# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Klard – Booking platform for consultants
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    get:
      operationId: listMyBookings
      tags: [bookings]
      summary: List bookings for the authenticated user (as customer)
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (confirm / cancel)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /providers/{id}/bookings:
    get:
      operationId: listProviderBookings
      tags: [bookings]
      summary: List bookings for a provider (dashboard)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
...
  /ai/offer:
    post:
      operationId: generateAiOffer
      tags: [ai]
      summary: Generate an AI offer for a service inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiOfferRequest"
      responses:
        "200":
          description: Generated offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiOfferResponse"
...
components:
  schemas:
...
    Booking:
      type: object
      required: [id, customerId, providerId, serviceId, slotId, status, totalPrice, scheduledAt]
      properties:
        id:           { type: integer }
        customerId:   { type: string }
        customerName: { type: ["string", "null"] }
        providerId:   { type: integer }
        providerName: { type: string }
        serviceId:    { type: integer }
        serviceName:  { type: string }
        slotId:       { type: integer }
        status:       { type: string, enum: [pending, confirmed, cancelled, completed] }
        totalPrice:   { type: number }
        scheduledAt:  { type: string, format: date-time }
        notes:        { type: ["string", "null"] }
        paymentRequired: { type: boolean }
        paymentStatus:   { type: string, enum: [not_required, pending, paid, failed, refunded] }
        stripeCheckoutSessionId: { type: ["string", "null"] }
        createdAt:    { type: string }
    BookingInput:
      type: object
      required: [providerId, serviceId, slotId]
      properties:
        providerId:  { type: integer }
        serviceId:   { type: integer }
        slotId:      { type: integer }
        notes:       { type: string }
    BookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [confirmed, cancelled, completed] }
...
    AiOfferRequest:
      type: object
      required: [providerId, inquiry]
      properties:
        providerId: { type: integer }
        inquiry:    { type: string }
    AiOfferResponse:
      type: object
      required: [offer]
      properties:
        offer:       { type: string }
        estimatedPrice: { type: ["number", "null"] }
        estimatedDuration: { type: ["string", "null"] }
...